+++ core-app/src/main/kotlin/com/mvcoding/expensius/feature/transaction/TransactionsPresenter.kt
@@ -18,21 +18,14 @@ import com.mvcoding.expensius.RxSchedulers
 import com.mvcoding.expensius.feature.ItemsView
 import com.mvcoding.expensius.feature.LoadingView
 import com.mvcoding.expensius.feature.ModelDisplayType
-import com.mvcoding.expensius.model.Money
-import com.mvcoding.expensius.model.NullModels.newTransaction
-import com.mvcoding.expensius.model.TimestampProvider
 import com.mvcoding.expensius.model.Transaction
-import com.mvcoding.expensius.service.AppUserService
 import com.mvcoding.expensius.service.TransactionsService
 import com.mvcoding.mvp.Presenter
 import rx.Observable
-import java.math.BigDecimal.ZERO
 
 class TransactionsPresenter(
         private val modelDisplayType: ModelDisplayType,
-        private val appUserService: AppUserService,
         private val transactionsService: TransactionsService,
-        private val timestampProvider: TimestampProvider,
         private val schedulers: RxSchedulers) : Presenter<TransactionsPresenter.View>() {
 
     override fun onViewAttached(view: View) {
@@ -68,12 +61,8 @@ class TransactionsPresenter(
                 .subscribeUntilDetached { view.showMovedItem(it.fromPosition, it.toPosition, it.item) }
 
         view.archivedTransactionsRequests().subscribeUntilDetached { view.displayArchivedTransactions() }
-
-        val newTransactions = view.createTransactionRequests()
-                .withLatestFrom(appUserService.appUser()) { unit, appUser ->
-                    newTransaction(timestampProvider.currentTimestamp(), Money(ZERO, appUser.settings.mainCurrency))
-                }
-        view.transactionSelects().mergeWith(newTransactions).subscribeUntilDetached { view.displayTransactionEdit(it) }
+        view.createTransactionRequests().subscribeUntilDetached { view.displayCalculator() }
+        view.transactionSelects().subscribeUntilDetached { view.displayTransactionEdit(it) }
     }
 
     interface View : Presenter.View, ItemsView<Transaction>, LoadingView {
@@ -84,6 +73,7 @@ class TransactionsPresenter(
         fun showModelDisplayType(modelDisplayType: ModelDisplayType)
 
         fun displayTransactionEdit(transaction: Transaction)
+        fun displayCalculator()
         fun displayArchivedTransactions()
     }
 }